dedupe decompression utils
